Refactor demos and add progress indicator
